# api/openapi/native-hosted-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Artifact Gateway Native Hosted Management API",
    "version": "1.0.0",
    "description": "Frozen contract for the planned Native Hosted Platform V3."
  },
  "servers": [
    {
      "url": "https://gateway.example.com/api/v2"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/audits": {
      "get": {
        "operationId": "listAudits",
        "description": "Lists administrative audit records. Authorization decision fields are present only when the record represents a repository authorization decision.",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repository",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "description": "Exact audit outcome filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Exact artifact format filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "operation",
            "in": "query",
            "description": "Exact HTTP or management operation filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "description": "Exact actor or principal filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuditList"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/repositories": {
      "get": {
        "operationId": "listRepositories",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/PageToken"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RepositoryList"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "operationId": "createRepository",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepository"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Repository"
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/repositories/{repositoryId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RepositoryId"
        }
      ],
      "get": {
        "operationId": "getRepository",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Repository"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "delete": {
        "operationId": "deleteRepository",
        "responses": {
          "202": {
            "$ref": "#/components/responses/Deletion"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "patch": {
        "operationId": "updateRepository",
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRepository"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Repository"
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "412": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/publish-sessions/{sessionId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        }
      ],
      "get": {
        "operationId": "getPublishSession",
        "responses": {
          "200": {
            "$ref": "#/components/responses/PublishSession"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/publish-sessions/{sessionId}:commit": {
      "parameters": [
        {
          "$ref": "#/components/parameters/SessionId"
        }
      ],
      "post": {
        "operationId": "commitPublishSession",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Artifact"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/raw/{repository}/": {
      "servers": [
        {
          "url": "https://gateway.example.com"
        }
      ],
      "parameters": [
        {
          "$ref": "#/components/parameters/RepositoryName"
        }
      ],
      "get": {
        "operationId": "listRawRootDirectory",
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "last",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "protocolBasicAuth": []
          },
          {
            "protocolBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RawDirectoryList"
          },
          "401": {
            "$ref": "#/components/responses/BasicUnauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRepository": {
        "type": "object",
        "required": [
          "name",
          "format"
        ],
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/Format"
          },
          "type": {
            "type": "string",
            "enum": [
              "hosted",
              "proxy"
            ],
            "default": "hosted"
          },
          "endpoint": {
            "type": "string",
            "description": "Upstream base URL. Required when type is proxy; must be empty for hosted repositories."
          },
          "allowedHosts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Hosts the proxy may egress to. Required for raw, conan, and npm proxies. npm tarball redirects are limited to this list."
          },
          "anonymousRead": {
            "type": "boolean",
            "default": false,
            "description": "Anonymous read policy. Defaults to false and permits unauthenticated protocol GET/HEAD where effective policy allows it."
          },
          "egressProxy": {
            "$ref": "#/components/schemas/EgressProxy"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "enum": [
              "invalid_request",
              "invalid_state",
              "invalid_page_token",
              "access_denied",
              "not_found",
              "version_conflict",
              "idempotency_conflict",
              "coordinate_exists",
              "session_closed",
              "digest_mismatch",
              "retention_protected",
              "security_policy_denied",
              "artifact_quarantined",
              "internal_error"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "RepositoryPage": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Repository"
            }
          },
          "nextPageToken": {
            "type": "string"
          }
        }
      },
...
      "AuditList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AuditRecord"
        }
      },
...
      "UpdateRepository": {
        "type": "object",
        "description": "Editable repository management policy and proxy configuration. Hosted repositories only accept anonymousRead updates; name, format, and type are immutable after creation.",
        "additionalProperties": false,
        "properties": {
          "endpoint": {
            "type": "string",
            "description": "Upstream base URL (https). Required for proxy repositories."
          },
          "allowedHosts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Hosts the proxy may egress to. Required for raw, conan, and npm proxies. npm tarball redirects are limited to this list."
          },
          "anonymousRead": {
            "type": "boolean",
            "description": "Anonymous read policy. Defaults to false and permits unauthenticated protocol GET/HEAD where effective policy allows it."
          },
          "egressProxy": {
            "$ref": "#/components/schemas/EgressProxy"
          }
        }
      },
...
    },
    "parameters": {
      "RepositoryId": {
        "name": "repositoryId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "SessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "RepositoryName": {
        "name": "repository",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
        }
      },
...
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      },
      "PageToken": {
        "name": "pageToken",
        "in": "query",
        "schema": {
          "type": "string"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Problem response",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "RawDirectoryList": {
        "description": "Lexicographically ordered Native Raw directory-prefix page",
        "headers": {
          "Link": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RawDirectoryPage"
            }
          }
        }
      },
...
      "BasicUnauthorized": {
        "description": "Native protocol Basic authentication required when anonymous read policy is disabled",
        "headers": {
          "WWW-Authenticate": {
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^Basic "
            }
          }
        }
      },
...
      "Repository": {
        "description": "Repository",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Repository"
            }
          }
        }
      },
      "RepositoryList": {
        "description": "Repository page",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RepositoryPage"
            }
          }
        }
      },
...
      "PublishSession": {
        "description": "Publish session",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PublishSession"
            }
          }
        }
      },
      "Artifact": {
        "description": "Artifact",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Artifact"
            }
          }
        }
      },
...
      "Deletion": {
        "description": "Accepted asynchronous deletion. The repository is immediately unavailable to protocol traffic and is finalized by the background deletion worker; audit metadata is retained.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Deletion"
            }
          }
        }
      },
      "AuditList": {
        "description": "Administrative audit records, ordered newest first",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuditList"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "protocolBasicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "protocolBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Gateway resolver token"
      },
      "ociBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OCI Registry token"
      }
    }
  },
...
}